Temporary

+++ gs2_time.py
@@ -16,7 +16,7 @@ class timeobj:
 
         # get starting index for steady-state
         self.twin = twin
-        self.it_min = ceil((1.0-twin)*self.ntime)
+        self.it_min = int(ceil((1.0-twin)*self.ntime))
         self.it_max = self.ntime-1
         # get number of time points in steady-state interval
         self.it_interval = self.ntime - self.it_min
